Always move the callback, see below for details
Even despite the usage of deduction guides trick, we cannot find out the value
category of callback constructor parameter, and assume it's always rvalue
(as std::declval returns an rvalue), hence trying to initialize the
std::function with an lvalue results in a compilation error (the deducted type
was for rvalue but we provided lvalue). So the solution is not to forward the
callback, but always move it.